Mover url resolver to rendermime namespace
